chore(core/option): break out `focusable` as reusable class mixin
